Split out types/constants
