+++ src/utils/notify.stories.tsx
@@ -18,7 +18,8 @@
 import React from "react";
 import { type ComponentStory, type ComponentMeta } from "@storybook/react";
 import { Button } from "react-bootstrap";
-import { initToaster, type NotificationType, showNotification } from "./notify";
+import { initToaster, showNotification } from "./notify";
+import { NotificationType } from "@/utils/notificationTypes";
 
 const notificationTypes = ["info", "success", "error", "warning", "loading"];
 initToaster();
